Splitting each class into its file Grouping files in several folder Refactor edge
