+++ include/kami/export/box_settings.hpp
@@ -1,0 +1,10 @@
+#ifndef KAMI_EXPORT_BOXES
+#define KAMI_EXPORT_BOXES
+
+namespace kami::out {
+
+constexpr double BOUNDS_PADDING{0.0f};
+
+}
+
+#endif
