+++ src/math/overlaps.cpp
@@ -1,4 +1,5 @@
-#include "kami/overlaps.hpp"
+#include "kami/math/overlaps.hpp"
+#include <algorithm>
 #include <ostream>
 
 namespace kami::overlaps {
